Alert display is converted to MVP pattern
